lab6/rlhf-preference-collector/requirements.txt
streamlit>=1.37.0
ollama>=0.3.0
supabase>=2.6.0
python-dotenv>=1.0.1
